addcard function to board
